+++ src/function/util.ts
@@ -8,7 +8,7 @@ export function substitudeMacros(text: string): string {
 }
 
 export function getLastMessageId(): number {
-  return chat.length - 1;
+  return Number(substitudeMacros('{{lastMessageId}}'));
 }
 
 export function errorCatched<T extends any[], U>(fn: (...args: T) => U): (...args: T) => U {